+++ src/extracts/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod file_watcher;
 pub mod fs;
+pub mod handlers;
 pub mod metrics;
 pub mod process_watcher;
 pub mod stdout;
